+++ src/Sulu/Component/Webspace/Tests/Unit/Analyzer/RequestAnalyzerTest.php
@@ -14,39 +14,83 @@ namespace Sulu\Component\Webspace\Tests\Unit\Analyzer;
 use Prophecy\Argument;
 use Sulu\Component\Webspace\Analyzer\Attributes\RequestAttributes;
 use Sulu\Component\Webspace\Analyzer\Attributes\RequestProcessorInterface;
+use Sulu\Component\Webspace\Analyzer\Exception\UrlMatchNotFoundException;
 use Sulu\Component\Webspace\Analyzer\RequestAnalyzer;
+use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
 class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
 {
-    public function testAnalyze()
+    public function testAnalyzeAndValidate()
     {
         $provider = $this->prophesize(RequestProcessorInterface::class);
-        $request = $this->prophesize(Request::class);
+        $request = new Request();
 
-        $provider->process($request->reveal(), Argument::type(RequestAttributes::class))
+        $provider->process($request, Argument::type(RequestAttributes::class))
             ->shouldBeCalled()->willReturn(new RequestAttributes());
         $provider->validate(Argument::type(RequestAttributes::class))->shouldBeCalled()->willReturn(true);
 
         $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal());
+        $requestStack->getCurrentRequest()->willReturn($request);
         $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
-        $requestAnalyzer->analyze($request->reveal());
+        $requestAnalyzer->analyze($request);
+        $requestAnalyzer->validate($request);
+    }
+
+    /**
+     * @expectedException \Sulu\Component\Webspace\Analyzer\Exception\UrlMatchNotFoundException
+     */
+    public function testAnalyzeAndValidateWithError()
+    {
+        $provider = $this->prophesize(RequestProcessorInterface::class);
+        $request = new Request();
+
+        $provider->process($request, Argument::type(RequestAttributes::class))
+            ->shouldBeCalled()->willReturn(new RequestAttributes());
+        $provider->validate(Argument::type(RequestAttributes::class))
+            ->shouldBeCalled()
+            ->willThrow(new UrlMatchNotFoundException(''));
+
+        $requestStack = $this->prophesize(RequestStack::class);
+        $requestStack->getCurrentRequest()->willReturn($request);
+        $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
+        $requestAnalyzer->analyze($request);
+        $requestAnalyzer->validate($request);
+    }
+
+    public function testAnalyzeWithoutValidateWithError()
+    {
+        $provider = $this->prophesize(RequestProcessorInterface::class);
+        $request = new Request();
+
+        $provider->process($request, Argument::type(RequestAttributes::class))
+            ->shouldBeCalled()->willReturn(new RequestAttributes());
+        $provider->validate(Argument::type(RequestAttributes::class))
+            ->shouldNotBeCalled()
+            ->willThrow(new UrlMatchNotFoundException(''));
+
+        $requestStack = $this->prophesize(RequestStack::class);
+        $requestStack->getCurrentRequest()->willReturn($request);
+        $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
+        $requestAnalyzer->analyze($request);
     }
 
     public function testGetAttribute()
     {
         $provider = $this->prophesize(RequestProcessorInterface::class);
-        $request = $this->prophesize(Request::class);
+        $request = new Request();
 
-        $provider->process($request->reveal(), Argument::type(RequestAttributes::class))
+        $provider->process($request, Argument::type(RequestAttributes::class))
             ->shouldBeCalledTimes(1)->willReturn(new RequestAttributes(['test' => 1]));
         $provider->validate(Argument::type(RequestAttributes::class))->shouldBeCalled()->willReturn(true);
 
         $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal());
+        $requestStack->getCurrentRequest()->willReturn($request);
         $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
+
+        $requestAnalyzer->analyze($request);
+        $requestAnalyzer->validate($request);
 
         $this->assertEquals(1, $requestAnalyzer->getAttribute('test'));
         $this->assertEquals(2, $requestAnalyzer->getAttribute('test1', 2));
@@ -55,35 +99,21 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
     public function testGetAttributeTwice()
     {
         $provider = $this->prophesize(RequestProcessorInterface::class);
-        $request = $this->prophesize(Request::class);
+        $request = new Request();
 
-        $provider->process($request->reveal(), Argument::type(RequestAttributes::class))
+        $provider->process($request, Argument::type(RequestAttributes::class))
             ->shouldBeCalledTimes(1)->willReturn(new RequestAttributes(['test' => 1]));
         $provider->validate(Argument::type(RequestAttributes::class))->shouldBeCalled()->willReturn(true);
 
         $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal());
+        $requestStack->getCurrentRequest()->willReturn($request);
         $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
+
+        $requestAnalyzer->analyze($request);
+        $requestAnalyzer->validate($request);
 
         $this->assertEquals(1, $requestAnalyzer->getAttribute('test'));
         $this->assertEquals(2, $requestAnalyzer->getAttribute('test1', 2));
-
-        $this->assertEquals(1, $requestAnalyzer->getAttribute('test'));
-        $this->assertEquals(2, $requestAnalyzer->getAttribute('test1', 2));
-    }
-
-    public function testGetAttributeBasic()
-    {
-        $provider = $this->prophesize(RequestProcessorInterface::class);
-        $request = $this->prophesize(Request::class);
-
-        $provider->process($request->reveal(), Argument::type(RequestAttributes::class))
-            ->shouldBeCalledTimes(1)->willReturn(new RequestAttributes(['test' => 1]));
-        $provider->validate(Argument::type(RequestAttributes::class))->shouldBeCalled()->willReturn(true);
-
-        $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal());
-        $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
 
         $this->assertEquals(1, $requestAnalyzer->getAttribute('test'));
         $this->assertEquals(2, $requestAnalyzer->getAttribute('test1', 2));
@@ -94,6 +124,16 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
         $request = $this->prophesize(Request::class);
         $request->getHost()->willReturn('www.sulu.io');
         $request->getScheme()->willReturn('https');
+
+        $attributesBag = $this->prophesize(ParameterBag::class);
+        $attributesBag->get('_sulu')->willReturn(null);
+        $attributesBag->set('_sulu', Argument::type(RequestAttributes::class))->shouldBeCalledTimes(1)->will(
+            function ($arguments) use ($attributesBag) {
+                $attributesBag->get('_sulu')->willReturn($arguments[1]);
+            }
+        );
+        $attributesBag->has('_sulu')->willReturn(true);
+        $request->reveal()->attributes = $attributesBag->reveal();
 
         $requestStack = $this->prophesize(RequestStack::class);
         $requestStack->getCurrentRequest()->willReturn($request->reveal());
@@ -143,15 +183,17 @@ class RequestAnalyzerTest extends \PHPUnit_Framework_TestCase
     public function testGetter(array $attributes, $method, $expected)
     {
         $provider = $this->prophesize(RequestProcessorInterface::class);
-        $request = $this->prophesize(Request::class);
-        $provider->process($request->reveal(), Argument::type(RequestAttributes::class))
+        $request = new Request();
+
+        $provider->process($request, Argument::type(RequestAttributes::class))
             ->shouldBeCalled()->willReturn(new RequestAttributes($attributes));
         $provider->validate(Argument::type(RequestAttributes::class))->shouldBeCalled()->willReturn(true);
 
         $requestStack = $this->prophesize(RequestStack::class);
-        $requestStack->getCurrentRequest()->willReturn($request->reveal());
+        $requestStack->getCurrentRequest()->willReturn($request);
         $requestAnalyzer = new RequestAnalyzer($requestStack->reveal(), [$provider->reveal()]);
-        $requestAnalyzer->analyze($request->reveal());
+        $requestAnalyzer->analyze($request);
+        $requestAnalyzer->validate($request);
 
         $this->assertEquals($expected, $requestAnalyzer->{$method}());
     }